# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Timeless Love API
  description: Backend API for Timeless Love family social platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@timelesslove.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.timelesslove.com
    description: Production server
...
paths:
  /api/v1/auth/register/adult:
    post:
      tags:
        - auth
      summary: Register Adult User
      description: Register a new adult user and create family unit
      operationId: registerAdult
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdultRegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/register/teen:
    post:
      tags:
        - auth
      summary: Register Teenager User
      description: Register a new teenager user
      operationId: registerTeen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeenRegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/auth/register/grandparent:
    post:
      tags:
        - auth
      summary: Register Grandparent User
      description: Register a new grandparent user
      operationId: registerGrandparent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrandparentRegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/auth/register/child:
    post:
      tags:
        - auth
      summary: Register Child User
      description: Register a child user via invitation token
      operationId: registerChild
      parameters:
        - name: invite_token
          in: query
          required: true
          schema:
            type: string
          description: Invitation token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChildRegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/auth/provision/child:
    post:
      tags:
        - auth
      summary: Provision Child Account
      description: Create a child account with generated credentials (Adult only)
      operationId: provisionChild
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChildProvisionRequest'
      responses:
        '201':
          description: Child account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildProvisionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/auth/register/pet:
    post:
      tags:
        - auth
      summary: Register Pet Profile
      description: Create a pet profile (Adult only)
      operationId: registerPet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetRegisterRequest'
      responses:
        '201':
          description: Pet profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetRegisterResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/auth/login:
    post:
      tags:
        - auth
      summary: Login
      description: Authenticate with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      tags:
        - auth
      summary: Refresh Token
      description: Refresh access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AdultRegisterRequest:
      type: object
      required:
        - email
        - password
        - display_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        display_name:
          type: string
          minLength: 1
          maxLength: 255
        family_name:
          type: string
          maxLength: 255
    TeenRegisterRequest:
      $ref: '#/components/schemas/AdultRegisterRequest'
    GrandparentRegisterRequest:
      $ref: '#/components/schemas/AdultRegisterRequest'
    ChildRegisterRequest:
      type: object
      required:
        - email
        - password
        - display_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        display_name:
          type: string
          minLength: 1
    ChildProvisionRequest:
      type: object
      required:
        - display_name
      properties:
        display_name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
    PetRegisterRequest:
      type: object
      required:
        - display_name
      properties:
        display_name:
          type: string
          minLength: 1
        avatar_url:
          type: string
          format: uri
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [adult, teen, child, grandparent, pet]
        family_unit_id:
          type: string
          format: uuid
        tokens:
          $ref: '#/components/schemas/TokenResponse'
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: Seconds until access token expires
    ChildProvisionResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        password:
          type: string
        display_name:
          type: string
        family_unit_id:
          type: string
          format: uuid
        message:
          type: string
    PetRegisterResponse:
      type: object
      properties:
        pet_id:
          type: string
          format: uuid
        display_name:
          type: string
        family_unit_id:
          type: string
          format: uuid
        message:
          type: string
...
    Error:
      type: object
      properties:
        detail:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'